Just implement createGenericContext() locally where it is used
